@@ -8,13 +8,15 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "mobx": "^6.12.3",
     "mobx-react": "^9.1.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
     "styled-components": "^6.1.11",
-    "typescript": "^4.4.4",
+    "typescript": "5.1.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -49,6 +51,7 @@
     "@babel/preset-typescript": "^7.24.1",
     "babel-loader": "^9.1.3",
     "copy-webpack-plugin": "^12.0.2",
+    "eslint-config-next": "^14.2.4",
     "html-webpack-plugin": "^5.6.0",
     "ts-loader": "^9.5.1",
     "webpack": "^5.91.0",
