@@ -9,8 +9,8 @@ const config = {
   ],
   "extends": [
     "next/core-web-vitals",
-    "plugin:@typescript-eslint/recommended-type-checked",
-    "plugin:@typescript-eslint/stylistic-type-checked"
+    // "plugin:@typescript-eslint/recommended-type-checked",
+    // "plugin:@typescript-eslint/stylistic-type-checked"
   ],
   "rules": {
     "@typescript-eslint/array-type": "off",
@@ -37,6 +37,7 @@ const config = {
         }
       }
     ]
-  }
+  },
+  "ignorePatterns": ["src/index.js", "src/App.js", "src/trpc/*.ts"],
 }
 module.exports = config;
