@@ -1,25 +1,25 @@
-import "~/styles/globals.css";
+// import "~/styles/globals.css";
 
-import { GeistSans } from "geist/font/sans";
+// import { GeistSans } from "geist/font/sans";
 
-import { TRPCReactProvider } from "~/trpc/react";
+// import { TRPCReactProvider } from "~/trpc/react";
 
-export const metadata = {
-  title: "Create T3 App",
-  description: "Generated by create-t3-app",
-  icons: [{ rel: "icon", url: "/favicon.ico" }],
-};
+// export const metadata = {
+//   title: "Create T3 App",
+//   description: "Generated by create-t3-app",
+//   icons: [{ rel: "icon", url: "/favicon.ico" }],
+// };
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en" className={`${GeistSans.variable}`}>
-      <body>
-        <TRPCReactProvider>{children}</TRPCReactProvider>
-      </body>
-    </html>
-  );
-}
+// export default function RootLayout({
+//   children,
+// }: {
+//   children: React.ReactNode;
+// }) {
+//   return (
+//     <html lang="en" className={`${GeistSans.variable}`}>
+//       <body>
+//         <TRPCReactProvider>{children}</TRPCReactProvider>
+//       </body>
+//     </html>
+//   );
+// }
